# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Boilerplate Service ",
        "version": "0.1"
    },
    "paths": {
        "/api/daily-volumes": {
            "post": {
                "tags": [
                    "Customer daily volumes"
                ],
                "summary": "Get a list of daily volumes with filters and pagination",
                "description": "Fetches a list of daily volumes, with optional filtering based on fields in the daily volumes table, and supports pagination.",
                "operationId": "78ffcb422566f71d0fb8317416c58496",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of results per page",
                        "schema": {
                            "type": "integer",
                            "example": 50
                        }
                    },
                    {
                        "name": "created_at_from",
                        "in": "query",
                        "description": "Filter by start date for created_at",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "created_at_to",
                        "in": "query",
                        "description": "Filter by end date for created_at",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "updated_at_from",
                        "in": "query",
                        "description": "Filter by start date for updated_at",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "updated_at_to",
                        "in": "query",
                        "description": "Filter by end date for updated_at",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "customer_id",
                        "in": "query",
                        "description": "Filter by customer ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "customer_site_id",
                        "in": "query",
                        "description": "Filter by customer ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "pagination": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "next_page_url": {
                                                    "type": "string",
                                                    "example": "http://example.com/api/daily-volumes?page=2",
                                                    "nullable": true
                                                },
                                                "prev_page_url": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 50
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 200
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 4
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/DailyVolume"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/daily-volumes/view/{id}": {
            "get": {
                "tags": [
                    "Daily Volumes"
                ],
                "summary": "Get details of a specific Daily Volume record",
                "description": "Fetches details of a specific Daily Volume record by ID.",
                "operationId": "4c207938ded7f72412bc0c0cb04c7f04",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Daily Volume record",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "": {
                                            "$ref": "#/components/schemas/DailyVolume"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/daily-volumes/{id}": {
            "delete": {
                "tags": [
                    "Daily Volumes"
                ],
                "summary": "Delete a specific Daily Volume record",
                "description": "Deletes a specific Daily Volume record by ID.",
                "operationId": "b4b1d271be388bea133292d95f0d1fd8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Daily Volume record",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Daily Volume record deleted successfully"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/gas-costs": {
            "post": {
                "tags": [
                    "Gas Costs"
                ],
                "summary": "Get a list of gas costs with filters and pagination",
                "description": "Fetches a list of gas costs, with optional filtering based on fields in the gas costs table, and supports pagination.",
                "operationId": "365c237750f50f29c358751e4c48450b",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of results per page",
                        "schema": {
                            "type": "integer",
                            "example": 50
                        }
                    },
                    {
                        "name": "created_at_from",
                        "in": "query",
                        "description": "Filter by start date for created_at",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "created_at_to",
                        "in": "query",
                        "description": "Filter by end date for created_at",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "pagination": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "next_page_url": {
                                                    "type": "string",
                                                    "example": "http://example.com/api/gas-costs?page=2",
                                                    "nullable": true
                                                },
                                                "prev_page_url": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 50
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 200
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 4
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/GasCost"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/gas-costs/view/{id}": {
            "get": {
                "tags": [
                    "Gas Costs"
                ],
                "summary": "Get details of a specific GasCost record",
                "description": "Fetches details of a specific GasCost record by ID.",
                "operationId": "693a3ae53ba096f1b2f4af697d8b07ce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the GasCost record",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "": {
                                            "$ref": "#/components/schemas/GasCost"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/gas-costs/delete/{id}": {
            "delete": {
                "tags": [
                    "Gas Costs"
                ],
                "summary": "Delete a specific GasCost record",
                "description": "Deletes a specific GasCost record by ID.",
                "operationId": "664d71a476b32a3252e51fe96023887b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the GasCost record",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Gas Cost record deleted successfully"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "DailyVolume": {
                "title": "Daily Volume",
                "properties": {
                    "id": {
                        "description": "ID of the existing Daily Volume",
                        "type": "integer",
                        "example": 1
                    },
                    "customer_id": {
                        "description": "ID of the customer",
                        "type": "integer",
                        "example": 1
                    },
                    "customer_site_id": {
                        "description": "ID of the customer site",
                        "type": "integer",
                        "example": 1
                    },
                    "abnormal_status": {
                        "description": "to indicate if the volume is normal or abnormal",
                        "type": "string",
                        "example": "normal"
                    },
                    "volume": {
                        "description": "Volume in Scf",
                        "type": "number",
                        "format": "float",
                        "example": 1
                    },
                    "created_at": {
                        "description": "Created at",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Updated at",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "GasCost": {
                "title": "Gas Cost",
                "properties": {
                    "id": {
                        "description": "ID of the gas cost record",
                        "type": "integer",
                        "example": 1
                    },
                    "date_of_entry": {
                        "description": "Date of entry",
                        "type": "string",
                        "format": "date",
                        "example": "2024-09-23"
                    },
                    "dollar_cost_per_scf": {
                        "description": "Dollar cost per SCF",
                        "type": "number",
                        "format": "float",
                        "example": 1
                    },
                    "dollar_rate": {
                        "description": "Dollar rate in NGN per dollar cost per SCF",
                        "type": "number",
                        "format": "float",
                        "example": 1
                    },
                    "status": {
                        "description": "Status of the gas cost record",
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "description": "Created at",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Updated at",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Customer daily volumes",
            "description": "Customer daily volumes"
        },
        {
            "name": "Daily Volumes",
            "description": "Daily Volumes"
        },
        {
            "name": "Gas Costs",
            "description": "Gas Costs"
        }
    ]
}